+++ src/core/host/host_prober_factory.hpp
@@ -1,0 +1,11 @@
+#pragma once
+
+#include "host_prober.hpp"
+
+#include <memory>
+
+namespace netfin::core::host {
+  namespace HostProberFactory {
+    std::unique_ptr<HostProber> create();
+  };
+}
