+++ src/core/platform.hpp
