+++ src/platform/apple/host/host_prober_apple.hpp
@@ -1,12 +1,12 @@
 #pragma once
 
-#include "network/host_prober/host_prober.hpp"
+#include "core/host/host_prober.hpp"
 #include <netinet/in.h>
 
-namespace netfin::network {
-  class HostProberLinux : public HostProber {
+namespace netfin::core::host {
+  class HostProberApple : public HostProber {
     public:
-      ~HostProberLinux() = default;
+      ~HostProberApple() = default;
 
       bool probe(
         std::string_view host,
@@ -14,17 +14,12 @@ namespace netfin::network {
       ) const override;
 
     private:
-      bool dgram_icmp_probe(
-        const sockaddr_in& addr, 
-        const std::chrono::milliseconds& timeout
-      ) const;
-
-      bool raw_icmp_probe(
-        const sockaddr_in& addr, 
-        const std::chrono::milliseconds& timeout
-      ) const;
-
       bool tcp_probe(
+        const sockaddr_in& addr, 
+        const std::chrono::milliseconds& timeout
+      ) const;
+
+      bool udp_probe(
         const sockaddr_in& addr, 
         const std::chrono::milliseconds& timeout
       ) const;
