+++ src/platform/linux/host/host_prober_linux.hpp
@@ -1,12 +1,12 @@
 #pragma once
 
-#include "network/host_prober/host_prober.hpp"
+#include "core/host/host_prober.hpp"
 #include <netinet/in.h>
 
-namespace netfin::network {
-  class HostProberApple : public HostProber {
+namespace netfin::core::host {
+  class HostProberLinux : public HostProber {
     public:
-      ~HostProberApple() = default;
+      ~HostProberLinux() = default;
 
       bool probe(
         std::string_view host,
@@ -14,12 +14,17 @@ namespace netfin::network {
       ) const override;
 
     private:
-      bool tcp_probe(
+      bool dgram_icmp_probe(
         const sockaddr_in& addr, 
         const std::chrono::milliseconds& timeout
       ) const;
 
-      bool udp_probe(
+      bool raw_icmp_probe(
+        const sockaddr_in& addr, 
+        const std::chrono::milliseconds& timeout
+      ) const;
+
+      bool tcp_probe(
         const sockaddr_in& addr, 
         const std::chrono::milliseconds& timeout
       ) const;
