+++ src/platform/windows/host_prober_windows.cpp
@@ -1,9 +1,9 @@
-#include "network/host_prober/host_prober_windows.hpp"
+#include "host_prober_windows.hpp"
 #include "core/platform.hpp"
 
 #if NETFIN_PLATFORM_WINDOWS
 
-namespace netfin::network {
+namespace netfin::core::host {
   bool HostProberWindows::probe(
     std::string_view host,
     const std::chrono::milliseconds& timeout
